+++ classes/set_dates_form.php
@@ -67,6 +67,7 @@ class set_dates_form extends \moodleform {
         // Don't overwrite existing enddates.
 
         $mform->addElement('checkbox', 'keepexisting', get_string('keepexistingenddates', 'tool_coursedates'));
+        $mform->setDefault('keepexisting', TOOL_COURSEDATES_KEEPEXISTING_NO);
 
         // Metadata.
         $mform->addElement('hidden', 'category');